requirements.txt
ml_collections
jax
jaxlib
flax
tensorflow
clu
tensorflow_transform
orbax-export    